checks: Rename wayland.py to linux.py
Groundwork for future linux-specific checks similar to the Windows and
MacOS specific checks.
