+++ loganalyzer.py
@@ -12,7 +12,7 @@ from checks.macos import *
 from checks.network import *
 from checks.plugins import *
 from checks.sources import *
-from checks.wayland import *
+from checks.linux import *
 from checks.windows import *
 
 from checks.utils.fetchers import *
